requirements.txt
botorch>=0.10.0,<0.12.0
cvxpy
scikit-learn
scipy
matplotlib
numpy
pre-commit
pytest
pytest-xdist
pytest-cov
bandit
black
usort
flake8
flake8-print
jupyter
pydocstyle
torchvision
optuna
fvcore
iopath

-r docs/requirements.txt
